refactor: AutoSyncConfig 리포지토리 패키지 구조 정비 및 잘못된 sortField 입력 시 기본값 폴백 처리
